fix: a length is checked before it narrows
`Inst::Alloc` with `Len::Slot` read an `i64` out of a slot and cast it
to `u32` with no check, and `payload_words` then multiplied the
narrowed count by the element stride in `u32`. A negative count wrapped
to something enormous; a count times a stride could wrap to something
small — an under-allocation, followed by writes sized by the original
logical length.

Nothing reaches it that way today: the lowering only ever passes counts
it computed. That is an invariant of one caller, not of the
instruction, and the bytecode verifier's own module doc says what the
standard is — "this must be safe against arbitrary bytes, because a
verifier that is only safe against its own encoder is not a verifier."

`Machine::allocate` takes an `i64` now and is the one checked entry
point: `u32::try_from` rejects negative and too-large in one call, then
`Layout::try_payload_words` does the same shape-driven arithmetic in
`u64` with `checked_mul` and `checked_add` before narrowing back. All
three failures land on the error a full heap already produces rather
than a new one, because from the machine's side a request it cannot
satisfy is not a different kind of failure.

`payload_words` itself is unchanged. It is called per live object
during a sweep, where the length comes from the object's own header and
cannot be untrusted, and widening it there would put a `u64` multiply
and a range check on the collector's path for no reason. The duplicated
rule is closed by a test asserting the two agree wherever both are
defined.

One boundary earlier than asked for: `alloc.imm`'s count is known
statically, so the bytecode verifier now refuses one whose payload
overflows before dispatch ever sees it. `Len::Slot` cannot be checked
that way — its value is a runtime slot — which is exactly why the
runtime check exists.

Four tests drive a real `Inst::Alloc` through the encoded dispatch loop
with a slot value the test controls: negative, past `u32::MAX`,
stride overflow, past the heap budget.

## On performance, which the issue asks about

I cannot say this is free and I cannot say it is not. A before/after
pair of `cove-bench` flagged eight rows, but `pure` and `call` — which
never allocate — moved as much as or more than the allocating rows,
which is the signature of layout noise rather than of this change.
`docs/VM_ARCHITECTURE.md` documents this same benchmark set swinging 1
to 11% from code layout alone, and settling it needs that document's
fifteen-sample protocol rather than one pair. Reported rather than
guessed at.

Verified: `cargo t`, `cargo ratchet` with the corpus unchanged,
clippy, `cargo doc`, `cove reference --check`, `cove fmt --check`, the
examples on both backends at 165 tests each, and `node web/check.mjs`.

+++ crates/cove-runtime/src/vm/exec/encoded.rs
@@ -887,10 +887,17 @@ pub(super) fn dispatch<'s, 'a>(
             // discriminant in a field, so nothing is stored and nothing is
             // asked.
             ALLOC_FIXED | ALLOC_IMM | ALLOC_SLOT => {
+                // `Len::Slot`'s word is read as `i64` and handed to
+                // `Machine::allocate` whole, not narrowed here: a negative
+                // count or one past what a `u32` header field can hold is
+                // that call's to reject, the same way it rejects
+                // `ALLOC_IMM`'s `Half::Count` — the other half `verify`
+                // does not range-check, for want of a table to check it
+                // against.
                 let len = match held.opcode() {
                     ALLOC_FIXED => 0,
-                    ALLOC_IMM => held.hi(),
-                    _ => machine.mem.slot(base, b!()) as u32,
+                    ALLOC_IMM => held.hi() as i64,
+                    _ => machine.mem.slot(base, b!()) as i64,
                 };
                 machine.sync(pc - 1);
                 match machine.allocate(LayoutId(held.lo()), len) {
@@ -1029,7 +1036,7 @@ pub(super) fn dispatch<'s, 'a>(
                 let layout = LayoutId(held.lo());
                 let width = machine.width(layout);
                 machine.sync(pc - 1);
-                let boxed = match machine.allocate(machine.boxed_layout(), width) {
+                let boxed = match machine.allocate(machine.boxed_layout(), width as i64) {
                     Ok(addr) => addr,
                     Err(error) => fail!(error),
                 };
